telegram_sidecar/bot/keyboard.py
from __future__ import annotations

from telegram import InlineKeyboardButton, InlineKeyboardMarkup


def home_keyboard() -> InlineKeyboardMarkup:
    return InlineKeyboardMarkup(
        [[InlineKeyboardButton("🩺 Engine", callback_data="v:engine")]]
    )
