requirements.txt
aiohttp>=3.9,<4.0
aiosqlite>=0.20,<1.0
python-dotenv>=1.0,<2.0
websockets>=12.0,<13.0
prometheus-client==0.21.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
python-telegram-bot>=21.0,<22.0
psutil>=6.0,<7.0
